pricing-deck-app/requirements.txt
Flask>=3.0,<4.0
python-pptx>=0.6.23,<1.0
requests>=2.31,<3.0
gunicorn>=21.2,<23.0
Pillow>=10.0,<12.0
